Feat: Restructure project into MVC-specific folders
This commit reorganizes both the backend and frontend directories to more
explicitly follow the Model-View-Controller (MVC) architectural pattern
by introducing dedicated folders for these components.

Backend Changes:
- Created `backend/controllers/` directory.
- Moved route handler logic from `backend/routes/*.js` files into
  newly created controller files within `backend/controllers/`
  (e.g., `applicationController.js`, `jobPostingController.js`).
- `backend/routes/*.js` files have been updated to import and delegate
  to these controller functions, making them lean route definitions.
- `backend/models/` directory remains the source for data models.

Frontend Changes (React - within `frontend/client/src/`):
- Created `views/` directory:
  - Moved `CompanyView.js` and `WorkerView.js` from `components/`
    to `views/`. These components act as the primary views.
- Created `controllers/` directory:
  - Moved custom hooks (`useCompanyForm.js`, `useWorkerData.js`)
    from `hooks/` to `controllers/`.
  - Renamed hooks to include 'Controller' suffix (e.g.,
    `useCompanyFormController.js`, `useWorkerDataController.js`) to
    clarify their role. These hooks manage state and logic for the views.
- Updated all import statements across `App.js`, view components,
  and controller hooks to reflect the new file locations and names.
- The `services/api.js` file remains in `services/`.
- Empty `frontend/client/src/components/` and
  `frontend/client/src/hooks/` directories have been removed.

This restructuring enhances the project's organization, improves
separation of concerns, and makes the MVC pattern more visually
apparent in the codebase.
